Removed util.rs. Working on tests.
